+++ src/utils/constants/index.ts
@@ -1,21 +1,30 @@
+// External imports
 import * as Icons from 'react-icons/ri';
+
+// Interfaces
 import { PresetSettings } from '../Interfaces/PluginPresets/Presets.interface';
 import { AppActiveState, AppIsShowState } from '../Interfaces/App.interface';
 import { IActivePresetSettings } from '../Interfaces/PluginSettings.interface';
 
-// Customize constant names for each different plugin.
+// Constants
+const POSSIBLE = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789abcdefghijklmnopqrstuvwxyz0123456789';
+
+// Plugin Configuration
 const PLUGIN_NAME = 'Plugin template'; // INSERT PLUGIN NAME
 const PLUGIN_ICON = Icons['RiOrganizationChart']; // Change PLUGIN Icon Name
 const PLUGIN_ID = 'main-custom-content'; // Insert Plugin ID
 
+// Table and Preset Defaults
 const TABLE_NAME = 'table_name';
 const DEFAULT_PRESET_NAME = 'Default Preset';
 
+// Default Select Option
 const DEFAULT_SELECT_OPTION = {
   value: '',
   label: '',
 };
 
+// Default Preset Settings
 const DEFAULT_PRESET_SETTINGS: PresetSettings = {
   shown_image_name: '',
   shown_title_name: '',
@@ -23,6 +32,7 @@ const DEFAULT_PRESET_SETTINGS: PresetSettings = {
   selectedView: DEFAULT_SELECT_OPTION,
 };
 
+// Default Plugin Data
 const DEFAULT_PLUGIN_DATA = {
   [PLUGIN_NAME]: PLUGIN_NAME,
   activePresetId: '0000',
@@ -36,8 +46,8 @@ const DEFAULT_PLUGIN_DATA = {
   ],
 };
 
+// Preset Handle Actions
 const PresetHandleAction = {
-  // would be better to have them as enum but it gives me an error in run time
   delete: 'delete',
   rename: 'rename',
   duplicate: 'duplicate',
@@ -45,14 +55,7 @@ const PresetHandleAction = {
   new: 'new',
 };
 
-// enum PresetHandleAction {
-//   Delete = 'delete',
-//   Rename = 'rename',
-//   Duplicate = 'duplicate',
-//   Edit = 'edit',
-//   New = 'new',
-// }
-
+// Initial App State
 const INITIAL_IS_SHOW_STATE: AppIsShowState = {
   isShowPlugin: true,
   isShowSettings: false,
@@ -68,13 +71,16 @@ const INITIAL_CURRENT_STATE: AppActiveState = {
   activePresetIdx: 0,
 };
 
+// Default Selected Preset
 const DEFAULT_SELECTED_PRESET: IActivePresetSettings = {
   activePresetId: '',
   selectedTable: DEFAULT_SELECT_OPTION,
   selectedView: DEFAULT_SELECT_OPTION,
 };
 
+// Exported Constants
 export {
+  POSSIBLE,
   PLUGIN_ICON,
   PLUGIN_NAME,
   PLUGIN_ID,
